+++ CHANGELOG.md
@@ -2,5 +2,10 @@
 All major and minor version changes will be documented in this file. Details of
 patch-level version changes can be found in [commit messages](../../commits/master).
 
+## 2020.0.1 - 2020/10/08
+- Use `metprint` if available
+- Bugfixes
+- Improvements - can use the main functionality if using as a module
+
 ## 2020 - 2020/10/07
 - First release